+++ src/id_translation/dio/default/__init__.py
@@ -1,13 +1,13 @@
 """Default :class:`~id_translation.dio.DataStructureIO` implementations."""
 
 from ._dict import DictIO
+from ._scalar import ScalarIO
 from ._sequence import SequenceIO
 from ._set import SetIO
-from ._single_value import SingleValueIO
 
 __all__ = [
     "DictIO",
+    "ScalarIO",
     "SequenceIO",
     "SetIO",
-    "SingleValueIO",
 ]